Issue #20: Changed get_next to get and created further functions
