+++ get/get_next_char.c
@@ -1,25 +1,25 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   get_next_line.c                                    :+:      :+:    :+:   */
+/*   get_next_char.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: mapandel <mapandel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/04 03:17:08 by mapandel          #+#    #+#             */
-/*   Updated: 2019/11/13 18:08:07 by mapandel         ###   ########.fr       */
+/*   Updated: 2019/12/01 15:46:07 by mapandel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "get_next.h"
+#include "get.h"
 
 /*
 **	get_next_line: get next line
 **		Reads the a line from a file descriptor int
 **		The system function read is called for a BUFF_SIZE number
 **			of characters at a time
-**		This char* line should be newline or EOF terminated
+**		This char* line could be char or EOF terminated
 **		Duplicates and stores it in a parameter pointer on char*
-**		The last standing newline character is not copied
+**		The last standing character searched is not copied
 **		The next function call searches the next line in the given file
 **		Returns a positive value if a line was read,
 **			a zero value if the file was read entirely
@@ -51,17 +51,17 @@ static t_line		*ft_init_list(t_line *list, const int fd)
 	return (list);
 }
 
-static int			ft_check_buf(char *buf, t_line *list, char **line)
+static int			ft_check_buf(char *buf, t_line *list, char **line, char c)
 {
 	char	*tmp;
 
 	tmp = ft_strdup(*line);
 	ft_strdel(line);
-	if (ft_strrchr(buf, '\n'))
+	if (ft_strrchr(buf, c))
 	{
-		*line = ft_strcjoin(tmp, buf, '\n');
-		list->save = ft_strsub(buf, (unsigned int)ft_strclen(buf,
-			'\n') + 1, ft_strlen(buf) - ft_strclen(buf, '\n'));
+		*line = ft_strcjoin(tmp, buf, c);
+		list->save = ft_strsub(buf, (unsigned int)ft_strclen(buf, c) + 1,
+			ft_strlen(buf) - ft_strclen(buf, c));
 		ft_strdel(&buf);
 		ft_strdel(&tmp);
 		return (1);
@@ -81,7 +81,7 @@ static int			ft_end_gnl(t_line *list, char **line, char *buf)
 	return (0);
 }
 
-int					get_next_line(const int fd, char **line)
+int					get_next_char(const int fd, char **line, char c)
 {
 	static t_line	*list = NULL;
 	char			*buf;
@@ -92,7 +92,7 @@ int					get_next_line(const int fd, char **line)
 		|| !(*line = ft_strnew(0))
 		|| !(list = ft_init_list(list, fd)))
 		return (-1);
-	if (list->save && ft_check_buf(list->save, list, line))
+	if (list->save && ft_check_buf(list->save, list, line, c))
 		return (1);
 	ft_strdel(&list->save);
 	if (!(buf = ft_strnew(BUFF_SIZE)))
@@ -100,7 +100,7 @@ int					get_next_line(const int fd, char **line)
 	while ((i = read(fd, buf, BUFF_SIZE)))
 	{
 		buf[i] = '\0';
-		if (ft_check_buf(buf, list, line))
+		if (ft_check_buf(buf, list, line, c))
 			return (1);
 		if (i < BUFF_SIZE)
 			break ;
