+++ includes/get.h
@@ -1,19 +1,19 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   get_next.h                                         :+:      :+:    :+:   */
+/*   get.h                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: mapandel <mapandel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/02/04 03:17:20 by mapandel          #+#    #+#             */
-/*   Updated: 2019/11/18 23:13:56 by mapandel         ###   ########.fr       */
+/*   Updated: 2019/12/01 16:09:33 by mapandel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef GET_NEXT_H
-# define GET_NEXT_H
+#ifndef GET_H
+# define GET_H
 
-# define BUFF_SIZE 10000
+# define BUFF_SIZE 4096
 
 # include <unistd.h>
 
@@ -35,7 +35,9 @@ typedef struct		s_line
 	char			*save;
 }					t_line;
 
+char				*get_file(int fd);
 int					get_next_char(const int fd, char **line, char c);
 int					get_next_line(const int fd, char **line);
+char				*get_stdin(void);
 
 #endif
